refactor: view 패키지로 이동
